build: added makefiles to build repo
Makefiles are intended to allow arbitrary directory structure
for this repo. It only supports building a STM32 firmware for
now. More rules can be added (e.g. build library or binaries)
in future to provide a more complete solutions that involves
STM32 firmware interacting with software stack on a host.