build: added zephyr base and tools
Building Zephyr apps is similar to building to kmod for
Linux kernel in that an app would use build system of
Zephyr "workspace" to compile which is then ready to be
loaded on target platform (microcontroller, qemu, etc.).

stm-main hosts zephyr apps (for multiple boards) which would
then uses the aforementioned Zephyr workspace to build.