build: improve building at a subdir level
- Current hacky method to look for product name in all sub-directories
to enumerate all the targets to call at $(TOPDIR) only works if the
product names are constant strings but not if they consist make vars.
For example, the method won't correctly enumerate target for followig:

ZEPHYR_APP := $(ZEPHYR_APP_NAME).$(BOARD)

More elegant solution is to instead include all Makefiles in the
sub-directories below and enumerate names organically. This method
still requires a "find -name Makefiles" and generally works. It,
however, won't honor if a sub-directory is present but isn't included
in the parent Makefile. Punting on that problem for now.

- Current method also doesn't enumerate multiple target names in a
Makefile. Added fix for that.