build: added support to build zephyr apps (#18)
- build: added zephyr base and tools
Building Zephyr apps is similar to building to kmod for
Linux kernel in that an app would use build system of
Zephyr "workspace" to compile which is then ready to be
loaded on target platform (microcontroller, qemu, etc.).

stm-main hosts zephyr apps (for multiple boards) which would
then use the aforementioned Zephyr workspace to build.

- Added rule to build a zephyr app using "west"
- hello_world: added an example zephyr app
- build: improve building at a subdir level
Current hacky method to look for product name in all sub-directories
to enumerate all the targets to call at $(TOPDIR) only works if the
product names are constant strings but not if they consist make vars.
For example, the method won't correctly enumerate target for followig:

ZEPHYR_APP := $(ZEPHYR_APP_NAME).$(BOARD)

More elegant solution is to instead include all Makefiles in the
sub-directories below and enumerate names organically. This method
still requires a "find -name Makefiles" and generally works. It,
however, won't honor if a sub-directory is present but isn't included
in the parent Makefile. Punting on that problem for now.

Current method also doesn't enumerate multiple target names in a
Makefile. Added fix for that.