build: allow same name for multiple products in different paths
Same name is also allowed in a path for products for different rule.
For example, C_LIB=C_BIN=hello is allowed for a <PATH> but multiple
C_BIN=hello is not.

This commit also adds "realpath" as dependency in buildenv which might
be an irritant when calling "buildenv targets" such as "make env" if
it's not present in the host's environment.