build: added support to format C/C++ files (#25)
The style for C++ (and other languages in future) is specified in
$TOPDIR/.clang-format with exception for C which is separately
specified in $TOPDIR/.clang-format-c.

If a directory has NO_FORMAT file present then all the files
in that directory and its sub-directories will skip the format
check.

- Added "make format" target to run clang-format on new+modified src files.
It always invokes format.sh from within buildenv shell to guarantee a version
for "clang-format". It's possible that host machine has a different version
in which case "clang-format" may yield completely differently results.

- Applied clang-format to $TOPDIR/{cmds,libs,firmware/zephyr} while skipping
some which are the directories where code is borrowed from the internet.