build: add version info to buildenv image (#27)
And print version info when there's a mismatch in currently running
buildenv docker image and version in build/Dockerfile.buildenv on
the current branch. That way it provides hint without impeding the
functionality. User should keep the image up to date for build to
work as expected.

- Added an unsavory measure to allow "docker" commands inside
buildenv-shell on OS X.